+++ leafs/userid/StaticUserIdLeaf.js
@@ -27,6 +27,7 @@ class StaticUserIdLeaf extends AbstractUserIdLeaf {
      * @inheritDoc
      *
      * @return {?string}
+     * @override
      */
     get userId() {
         return this._userId;
@@ -36,8 +37,11 @@ class StaticUserIdLeaf extends AbstractUserIdLeaf {
      * @inheritDoc
      *
      * @param {?string} userId
+     * @override
      */
     set userId(userId) {
+        this._commentOnUserIdType(userId);
+
         this._userId = userId;
     }